+++ app/assets/javascripts/components/CardApplicationSurvey/CardApplicationSurvey.js.jsx
@@ -1,4 +1,7 @@
 const React = require("react");
+const _     = require("underscore");
+const humps = require("humps");
+const $     = require("jquery");
 
 const ApplyActions     = require("./ApplyActions");
 const CallActions      = require("./CallActions");
@@ -12,16 +15,22 @@ const CardApplicationSurvey = React.createClass({
     applyPath:   React.PropTypes.string.isRequired,
     cardAccount: React.PropTypes.object.isRequired,
     declinePath: React.PropTypes.string.isRequired,
-    updatePath:  React.PropTypes.string.isRequired,
   },
+
+  getInitialState() {
+    return {
+      cardAccount: this.props.cardAccount,
+    };
+  },
+
 
   // Return the component that contains the actions for this card account.
   // Note that for some card accounts, no further action is required.
   getActionsComponent() {
-    const cardAccount = this.props.cardAccount;
+    const cardAccount = this.state.cardAccount;
 
     if (cardAccount.openedAt || cardAccount.redeniedAt) {
-      return undefined;
+      return "noscript";
     }
 
     if (!cardAccount.appliedAt) {
@@ -35,14 +44,37 @@ const CardApplicationSurvey = React.createClass({
     }
   },
 
+
+  submitAction(action, extraData) {
+    const data = {
+      _method: "patch",
+      "card_account[action]" : action,
+    };
+
+    if (action === "open" && extraData && extraData.openedAt) {
+      data["card_account[opened_at]"] = extraData.openedAt;
+    }
+
+    $.post(
+      `/api/v1/card_recommendations/${this.props.cardAccount.id}`,
+      data,
+      (newCardAccountAttrs, textStatus, jqXHR) => {
+        const oldCardAccount = this.state.cardAccount;
+        const newAttrs       = humps.camelizeKeys(newCardAccountAttrs);
+        this.setState({ cardAccount: _.assign(oldCardAccount, newAttrs) });
+      }
+    );
+  },
+
+
   render() {
     return React.createElement(
       this.getActionsComponent(),
       {
-        applyPath:   this.props.applyPath,
-        cardAccount: this.props.cardAccount,
-        declinePath: this.props.declinePath,
-        updatePath:  this.props.updatePath
+        applyPath:    this.props.applyPath,
+        cardAccount:  this.props.cardAccount,
+        declinePath:  this.props.declinePath,
+        submitAction: this.submitAction,
       }
     );
   },