+++ app/assets/javascripts/components/CardApplicationSurvey/ICalledButton.js.jsx
@@ -1,0 +1,17 @@
+const React = require("react");
+
+const Button = require("../core/Button");
+
+const ICalledButton = ({bankName, onClick}) => {
+  return (
+    <Button
+      onClick={onClick}
+      primary
+      small
+    >
+      I called {bankName}
+    </Button>
+  );
+};
+
+module.exports = ICalledButton;