+++ app/assets/javascripts/components/CardApplicationSurvey/IHeardBackButton.js.jsx
@@ -1,0 +1,28 @@
+const React = require("react");
+
+const Button = require("../core/Button");
+
+const IHeardBackButton = ({bankName, onClick}) => {
+  let text;
+  if (bankName) {
+    text = `I heard back from ${bankName} by mail or email`;
+  } else {
+    text = "I heard back from the bank";
+  }
+
+  const primary = typeof bankName === "undefined";
+
+  return (
+    <Button
+      default={!primary}
+      onClick={onClick}
+      primary={primary}
+      small
+    >
+      {text}
+    </Button>
+  );
+};
+
+module.exports = IHeardBackButton;
+