+++ app/assets/javascripts/components/CardApplicationSurvey/PostNudgeActions.js.jsx
@@ -1,99 +1,111 @@
 const React = require("react");
 
-const Button         = require("../core/Button");
-const ButtonGroup    = require("../core/ButtonGroup");
-const Form           = require("../core/Form");
-const HiddenFieldTag = require("../core/HiddenFieldTag");
-
 const ConfirmOrCancelBtns = require("../ConfirmOrCancelBtns");
 
 const ApprovedDeniedPendingBtnGroup = require("./ApprovedDeniedPendingBtnGroup");
+const IHeardBackButton              = require("./IHeardBackButton");
 
 const PostNudgeActions = React.createClass({
   propTypes: {
-    updatePath: React.PropTypes.string.isRequired,
+    submitAction: React.PropTypes.func.isRequired,
   },
 
 
   getInitialState() {
-    // Possible currentActions:
+    // Possible currentStates:
     // - initial
     // - heardBack
     // - confirmApproved
     // - confirmDenied
-    return { currentAction: "initial" };
+    return { currentState: "initial" };
   },
 
 
-  setCurrentAction(e, action) {
-    e.preventDefault();
-    this.setState({currentAction: action});
+  getAction() {
+    switch (this.state.currentState) {
+      case "confirmApproved":
+        return "open";
+      case "confirmDenied":
+        return "deny";
+      default:
+        throw "this should never happen";
+    }
+  },
+
+
+  getHelpText() {
+    switch (this.state.currentState) {
+      case "initial":
+        return "Let us know when you hear back from the bank:";
+      case "heardBack":
+        return "What did the bank say?";
+      case "confirmApproved":
+        return "Your application has been approved:";
+      case "confirmDenied":
+        return "Your application has been declined:";
+    }
+  },
+
+
+  setStateToApproved() {
+    this.setState({currentState: "confirmApproved"});
+  },
+
+
+  setStateToDenied() {
+    this.setState({currentState: "confirmDenied"});
+  },
+
+
+  setStateToInitial() {
+    this.setState({currentState: "initial"});
+  },
+
+
+  setStateToHeardBack() {
+    this.setState({currentState: "heardBack"});
+  },
+
+
+  submitAction() {
+    this.props.submitAction(this.getAction());
   },
 
 
   render() {
-    var buttons, helpText;
-
-    var action = "";
-
-    switch (this.state.currentAction) {
-      case "initial":
-        helpText = "Let us know when you hear back from the bank:"
-        break;
-      case "heardBack":
-        helpText = "What did the bank say?"
-        break;
-      case "confirmApproved":
-        helpText = "Your application has been approved:"
-        action = "open"
-        break;
-      case "confirmDenied":
-        helpText = "Your application has been declined:"
-        action = "deny"
-        break;
-    }
-
-    switch (this.state.currentAction) {
-      case "initial":
-        buttons = (
-          <Button
-            primary
-            small
-            onClick={e => this.setCurrentAction(e, "heardBack") }
-          >
-            I heard back from the bank
-          </Button>
-        );
-        break;
-      case "heardBack":
-        buttons = (
-          <ApprovedDeniedPendingBtnGroup
-            approvedText="My application was approved"
-            deniedText="My application was declined"
-            onClickApproved={e => this.setCurrentAction(e, "confirmApproved")}
-            onClickDenied={e => this.setCurrentAction(e, "confirmDenied")}
-          />
-        );
-        break;
-      case "confirmApproved":
-      case "confirmDenied":
-        buttons = (
-          <ConfirmOrCancelBtns
-            small
-            onClickCancel={e => this.setCurrentAction(e, "heardBack")}
-          />
-        );
-        break;
-    }
-
     return (
-      <Form action={this.props.updatePath} method="patch">
-        <HiddenFieldTag name="card_account[action]" value={action} />
+      <div>
+        <p>{this.getHelpText()}</p>
 
-        <p>{helpText}</p>
-
-        {buttons}
-      </Form>
+        {(() => {
+          switch (this.state.currentState) {
+            case "initial":
+              return (
+                <IHeardBackButton
+                  onClick={this.setStateToHeardBack}
+                />
+              );
+            case "heardBack":
+              return (
+                <ApprovedDeniedPendingBtnGroup
+                  approvedText="My application was approved"
+                  deniedText="My application was declined"
+                  onClickApproved={this.setStateToApproved}
+                  onClickDenied={this.setStateToDenied}
+                />
+              );
+            case "confirmApproved":
+            case "confirmDenied":
+              return (
+                <ConfirmOrCancelBtns
+                  onClickCancel={this.setStateToHeardBack}
+                  onClickConfirm={this.submitAction}
+                  small
+                />
+              );
+          }
+        })()}
+      </div>
     );
   },
 });